online_inference/requirements.txt
fastapi==0.65.1
uvicorn==0.13.4
pandas==1.2.1
PyYAML==5.4.1
scikit-learn==0.24.2
numpy==1.20.3